requirements.txt
python-telegram-bot==11.1.0
paramiko==2.4.2
pytesseract==0.2.7
